🔢 Keep VERSION file with each subgem
Storing the VERSION file only t the root breaks when the subgems are compiled
as the parent folder is no longer acessible. Soft linking the parent dir's
VERSION file to the two sub gem folders means that - via git - the file is
always stored in the compiled gem files and is available even when only
one of the gems is installed.